% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_mean_median_summarise.R
\name{produce_mean_median_summarise}
\alias{produce_mean_median_summarise}
\title{Produce Mean Median Summarise}
\usage{
produce_mean_median_summarise(
  .data,
  col_label,
  target_summarise_col_label,
  na.rm = TRUE
)
}
\arguments{
\item{.data}{the data (has to be in tibble version)}

\item{col_label}{the data column will be used for grouping.}

\item{target_summarise_col_label}{is the numeric data column will be used to calculate the statistic values.}

\item{na.rm}{is a bool variable will be used to drop na values or not.}
}
\value{
The mean and median target_summarise_col_label of the data grouped by col_label
}
\description{
Produce Mean Median Summarise
}
\details{
the function is designed for calculating the summary stats (mean and median) of given target stats label and group label.
}
\examples{
produce_mean_median_summarise(palmerpenguins::penguins, island, bill_depth_mm, TRUE)
produce_mean_median_summarise(palmerpenguins::penguins, island, bill_length_mm, TRUE)
}
